Handle git send-mail info line correctly
This code line appended to the lines of the chunk is used by git to
provide information such as what function or struct we're in. This is
NOT the first line of the context, as was previously supposed by the
add-on.

Handle this information line on its own: parse it and add it to the
different views. Prevent this line to appear in line numbers
recalculation.

Also try to amend tests, but I still don't have the setup to run them so
maybe I just messed up on this side.

This should fix GitHub issue #87.