transformations: fix usage of replace() to show all whitespaces on line
replace(/pattern/flags, "replacement") must be used instead of
replace("pattern", "replacement", "flags"). Flags as a third argument
was a Gecko extension, but is has been deprecated and support was
completely dropped for versions 49 and higher.

In particular, this update makes the global flag (g) work again in
transformations. The most visible effect seems to be the display of
whitespaces, which was broken and would only occur on the previous
occurrence of spaces and tabs in each line.

However, be careful not to replace all spaces with this transformation:
we do not want to remove spaces within HTML tags, it would invalidate
HTML markup and remove links on URLs for instance. A simple workaround
is used to avoid replacing within substrings enclosed in angle brackets.
It might not be really robust, but hopefully it should do the trick in
simple situations.

This commit fixes GitHub issue #86.