bump version number to 0.8
Minor code change, but we changed the license! Increase version number.

Also increase version number of highest supported Thunderbird version to
60. Note that Thunderbird version 61 should deprecate XUL-based
extension and require a major rewriting of the add-on.

Known issue: settings for colors are broken. Selection window is broken
on TB 59, and I did not even find a way to make it appear on TB 60. The
add-on comes with a fixed choice of colors, unless users want to change
them in the editor config.