Prevent coloring inside block quotes
I cannot find a reasonable solution for identifying diff snippets in
reply emails. Users answering online are susceptible to split diff
chunks. The add-on tries to strip all HTML to find the diffs, but the
result is often terrible and ends up mixing the messages of several
people, this is absolutely impossible to read.

One solution could consist in identifying all block quotes and trying to
find diffs in their contents, but apparently the add-on is not really
organized like this. It searches just one text for diff chunks, and
replace the whole DOM of the message by the one colored text it built
instead. I don't have the courage to rewrite it.

As a trade-off, simply refuse to strip HTML on block quotes. We are
better with no color than with completely broken email structure. And
if the parser fails to find diffs because of that, make sure we do not
display the toolbar.

This can be seen as a workaround for GitHub issue #65.