fix compatibility issues for Thunderbird 57~60
Fixed a number of issues that prevented the add-on to work correctly on
Thunderbird 57+.

See this page about the changes:
https://wiki.mozilla.org/Thunderbird/Add-ons_Guide_57

More specifically:

* Replace nsIPrefBranchInternal by nsIPerfBranch.
* Remove deprecated `for each ... in ... process(val)` loops and replace
  them by `for (... in ...) process(object[valName])`.
* Change type for scripts inclusions: `application/javascript` instead
  of `application/x-javascript`.

Note that this is work in progress!! Most of the add-on works on
Thunderbird 59 with this patch, however the popup used to set
preferences is badly broken. On Thunderbird 60, I could not even find a
way to make the popup appear.