Replace XBL element by simple <menuitem> for view mode
As far as I understand, the XBL bindings for the view mode are used so
as to avoid duplicating the code for the view menu list in the toolbar
and in the settings window.

Since XBL bindings are being deprecated [1], and the settings dialog is
broken anyway, replace it with a simple <menuitem><menupopup></></>
element.

This, along with previous commit, restores the proper behaviour for the
toolbar.

[1] https://developer.thunderbird.net/add-ons/tb68/changes#removed-xbl-bindings<Paste>