Remove (from toolbar) buttons to show options and close toolbar
Because the window for choosing the settings is completely broken at the
moment, comment out the button supposed to open that window.

Because we don't have a window for choosing the settings, remove the
button to close the toolbar. The only way to reactivate the toolbar is
to use the config editor, which is painful if user accidentally closes
the toolbar.