Remove broken settings page
The page for selecting the options for the add-on is completely broken
at the moment. We already removed the button to open it in a previous
commit; also remove the possibility to open it through Thunderbird
add-ons menu.

The only way to change the settings at the moment is through the config
editor.