v2: Port to TB 78.4+/82+, rewrite add-on from scratch, use highlight.js
Following Thunderbird's switch to MailExtensions, the add-on had to be
ported as a "legacy" extension for v68. This proved to be a short
relief, since the following stable, v78, does not support those anymore.

The new version of the add-on is a pure MailExtension that should be
compatible with new Thunderbird versions. It relies on the
messagesModify API that appeared in Thunderbird 82, but has since been
backported to version 78.4 [0].

Various parts of the add-on were referring to XUL overlays or chrome
components, but none are available with MailExtensions. Instead of
struggling to adapt these parts, the add-on was entirely rewritten from
scratch, and with a different approach. We no longer parse and reformat
the diffs: Instead, we embed and inject the highlight.js library that
takes care of coloring the diff [1].

Consequences and main changes:

- We loose some accuracy, because we do not rely on "what part of the
  diff we are in". Concretely, all lines starting with "+" or "-" will
  be colored, even if in the commit log.

- We gain in stability, with less parsing susceptible to break.

- We loose the different modes, unified / context / side-by-side. The
  diffs are not rewritten, we simply add colors.

- We keep the ability to replace tabs and spaces, but loose the line
  numbers.

- It is not possible at this time to specify custom colors for the
  diffs, but we can pick a style amongst the hundred or so that are
  supported by highlight.js.

Bumping the major version number.

[0] https://bugzilla.mozilla.org/show_bug.cgi?id=1504475
[1] https://highlightjs.org/