Fix list and checks for CSS styles
We had a Makefile target to check that all styles under hljs/styles/ are
listed in the options/fill-css-list.js file. But so far we wouldn't
check the opposite: whether all files listed in the code are indeed
present among the CSS files extracted from the library. Add a target to
check this.

We also need a fix for the existing check: Some CSS files were moved to
hljs/styles/base16/, and we need to check there as well.

Update the list of styles accordingly.