Create a "transformations" object to mark exports
This is in preparation of using JavaScript modules in the add-on. There
are a few cases where we are not going to use modules, because I believe
it is not possible to use them in content scripts.

This means we keep a few variables/functions defined globally, and made
available to the scope by loading several scripts. To make it clearer
that such functions are defined in a specific file and coming from a
global object, let's move the functions defined in transformations.js
into a dedicated object. This should also simplify imports/exports if
they become possible in the future.