Add Makefile target for Mozilla's add-on linter
This add-on linter is run on new submissions to AMO or ATN. Let's add a
Makefile target to help run it locally.

The linter first needs to be installed with:

    $ npm install addons-linter

I don't think it is necessary to make it a GitHub action, given that it
raises a few warnings that are due to a bad support for Thunderbird's
add=ons (Firefox is the first-class citizen) or some that we simply
won't fix (innerHTML assignments done by highlight.js or even the add-on
itself). It should be enough to run it before submitting a new version,
to make sure no error crept in.