Use hljs's cssSelector to colour blocks, skip links
Highlight.js has a cssSelector option to mark what block should be
highlighted. Instead of colouring all child elements of the <pre>
blocks, we can add a dedicated class name to all child text elements
(only) and use the option to only colour those.

The motivation for this change is to avoid colouring the links present
in the email body, to avoid removing the <a> mark up and keep the
possibility to click them (plus the style), as provided by Thunderbird.

The Git email signature is not plain text and hence is now skipped, so
we can remove the dedicated steps.

Fixes: #104
Co-developed-by: Leonardo Garcia <leonardo.garcia@linaro.org>