Fix whitespace display, bump patch version number
In a (more or less) recent commit, we changed the way we're building the
nodes for adding the colours. This introduced a bug for displaying white
spaces, because we didn't reflect the changes when passing the root node
to process to the relevant function.

This went undetected for a while because I didn't immediately publish a
new release with the buggy commit. Let's fix it now, and cut a patch
release with the bug fix.

Fixes: 2c2be9319d28 ("Use hljs's cssSelector to colour blocks, skip links")